fix(CollectionsController): Fixes zero length files in archive
Some blueprints have non latin charcters in their titles. This broke saving the contents to a temp file (if I have to guess). This fixes that issue by not using a temp file at all for saving the blueprint data to. Instead it writes the content of a blueprint to the archive in memory, speeding things up too.

Fixes #22