fix: base game manger now handles new patch notes
The DSP team has started putting a V before the numbers in their release
notes which breaks the base game manager's parsing of the patch notes
for new versions. This commit extends the regexp to handle the presence
(or lack thereof) of the V in the patch notes, as they've also been very
inconsistent.